tests/e2e/pyproject.toml
[project]
name = "reconfigurable-device-e2e"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["pytest>=8.0", "requests>=2.31", "websockets>=13.0", "zeroconf>=0.131", "pyserial>=3.5"]

[tool.pytest.ini_options]
testpaths = ["."]
addopts = "-x -v"
markers = [
    "reboot: tests that perform a hardware reboot via espflash",
    "stress: long-running stress and stability tests",
    "provisioning: WiFi provisioning flow tests (portal, form submit, reconnect)",
    "mdns: mDNS and DNS-SD discovery tests",
    "ota: OTA firmware update tests (requires OTA_FIRMWARE_A_GZ and OTA_FIRMWARE_B_GZ)",
    "wsop: WSOP secure onboarding tests (requires two devices: gateway + joiner)",
]
